Session 17: readability pass, C23 adoption, skew verification
Large naming and documentation cleanup across the BIOS source. Both
compilers build warning-free. Sizes unchanged: clang 6021, sdcc 6124.
MAME boot on SW1711-I8 verified (2685 sectors, 0 errors).

Compiler configuration:
- clang: -std=c23 adopted; 5 warnings fixed at source
  (imprecision cast at bios.c:517, 4× int-to-pointer in hal.h macros
  via uintptr_t widening)
- SDCC: warnings 147 (excess-init boot signature) and 221 (inline
  __naked ISR helpers) disabled; both are intentional patterns

FDC register vocabulary (bios.h):
- FDC_MSR_RQM/DIO/NDM/CB/DRIVE_SEEKING for Main Status Register
- FDC_ST0_INTERRUPT_CODE/IC_*/SEEK_END/EQUIPMENT_CHECK/NOT_READY
- FDC_ST0_ANY_ERROR composite
- Replaces 9 raw binary masks in bios.c + bios_hw_init.c
- fdc_read_result() now has comprehensive per-command doc + full
  ST0/ST1/ST2/ST3 bit tables from the uPD765 datasheet

Major variable rename (25+ identifiers):
- CP/M BDOS state: sekdsk/sektrk/seksec → cpm_disk/track/sector
- Host buffer: hstdsk/hsttrk/hstsec → hostbuf_disk/track/sector
- Seek cache: lstdsk/lsttrk → last_seek_disk/track
- Unallocated: unacnt/unadsk/unatrk/unasec/unamsk → unalloc_*
- I/O flags: rsflag → need_pre_read, readop → is_read, wrtype stays
- DMA/format: dmaadr → cpm_dma_addr, cform → current_format_idx,
  eotv → track0_sectors_per_side, drno → max_drive_num
- Physical/FDC: dskno/dskad/actra/acsec → fdc_unit_head/dma_addr/
  track/sector
- ISR helpers: fl_flg/clfit/watir/wfitr → fdc_irq_fired/arm/wait/
  wait_rearm
- FDC result: rstab → fdc_result; function renamed fdc_result() →
  fdc_read_result() to resolve type/function collision
- FSPA working-copy fields renamed to match struct field rename:
  dpblck → current_dpb, cpmrbp → records_per_alloc_block,
  cpmspt → cpm_sectors_per_track, secmsk/secshf → deblock_mask/
  deblock_shift, trantb → sector_xlate, dtlv → fdc_data_length,
  dsktyp → is_hard_disk

Type renames:
- DPB/DPB_s → disk_parameter_block/disk_parameter_block_s
- DPH → disk_parameter_header
- dpbase → dph_table
- Per-format data: tran0/8/16/24 → xlt_maxi_128/512/mini_512/
  xlt_identity (DRI XLT tradition); dpb0/8/16/24 → dpb_maxi_128/
  512/mini_512/maxi_256
- .dpb field kept short inside DPH/FSPA (CP/M idiom, avoids
  type/field name collision)

Code refactoring:
- sec_rw(): goto fail replaced with for-loop + break; static byte
  repet preserved inside function for BSS layout compatibility
- Halfway-recalibrate check: repet==5 (post-dec) → repet==6 (pre-dec)
- bios_conout_c(): default-case added for exhaustive 2-bit IOBYTE
- bios_seldsk_c(): return 0 → return nullptr (C23)
- Local variable conflicts from DPB/DPH typedef renames fixed by
  using `entry` for DPH pointer locals in bios.c and bios_hw_init.c

Build-time verification:
- verify_skew.py: parses xlt_*[] tables from bios.c, compares to
  output of the CP/M skew algorithm for each (SPT, skew) pair.
  Analog of the DRI assembler XLT macro, run as a Makefile
  prerequisite of the `bios` target. Build fails if any table
  drifts from spec.

Documentation:
- bios.c BSS declarations now column-aligned with trailing comments
- DPH and FSPA struct fields column-aligned in bios.h
- Skew tables reformatted: right-aligned 2-digit numbers, consistent
  spt/skew documentation per table

CP/M source material (cpm/):
- OS2CCP.ASM, OS3BDOS.ASM — authoritative DRI CP/M 2.2 sources
- rc700.z3t, rc700gr.z3t — original ZCPR3 TCAP binaries
- rc700_z3termcap.h — annotated C source (49 bytes, well-formed,
  Z-System extension bytes preserved with trailing null). Future
  ZCPR3 integration can #include this.

Follow-ups recorded in tasks/todo.md:
- FSPA initializer: consider designated initializers now that
  fields have meaningful names
- Remaining short-named vars (erflag, hstact, hstwrt, prtflg,
  ptpflg, cerflg, delcnt) — audit for rename
- rsflag semantic verification (confirm "pre-read required")
- DRI CCP/BDOS build from source (currently only BIOS is built)
- TCAP layout verification for ZCPR3 integration

+++ rcbios-in-c/hal.h
@@ -99,17 +99,21 @@
  * selected at runtime). Requires #44 fix on clang side. */
 #if defined(__clang__) && defined(__z80__)
 #define __io __attribute__((address_space(2)))
+/* Z80 I/O port numbers are 8-bit but pointers are 16-bit. The (uint8_t) cast
+ * enforces the 8-bit port-number semantic; (uintptr_t) then widens to the
+ * pointer-sized integer so the final cast is same-width and silences
+ * -Wint-to-pointer-cast. Both casts are no-ops in emitted Z80 code. */
 #define DEFPORT(name, addr) \
     static inline uint8_t port_in_##name(void) { \
-        return *(volatile __io uint8_t *)(uint8_t)(addr); \
+        return *(volatile __io uint8_t *)(uintptr_t)(uint8_t)(addr); \
     } \
     static inline void port_out_##name(uint8_t val) { \
-        *(volatile __io uint8_t *)(uint8_t)(addr) = val; \
+        *(volatile __io uint8_t *)(uintptr_t)(uint8_t)(addr) = val; \
     }
 #define port_in(name)       port_in_##name()
 #define port_out(name, val) port_out_##name(val)
-#define port_in_rt(p)       (*(volatile __io uint8_t *)(uint8_t)(p))
-#define port_out_rt(p, v)   (*(volatile __io uint8_t *)(uint8_t)(p) = (v))
+#define port_in_rt(p)       (*(volatile __io uint8_t *)(uintptr_t)(uint8_t)(p))
+#define port_out_rt(p, v)   (*(volatile __io uint8_t *)(uintptr_t)(uint8_t)(p) = (v))
 #elif defined(__SDCC) || defined(__SCCZ80)
 #define DEFPORT(name, addr) __sfr __at (addr) _sfr_##name;
 #define port_in(name)       (_sfr_##name)
